deprecate the js lambda codes
