--- src/lambda-assets/ca-registrator/certificate-generator.js
@@ -1,127 +1,0 @@
-'use strict';
-const forge = require('node-forge');
-
-exports.CertificateGenerator = class CertificateGenerator {
-  /**
-     * Generate a certificate template which can be further
-     * used to generate a CA or a verification certificate.
-     * @param {array} attr The formatted CSR subject.
-     * @param {Number} years The valid time interval of the generated certificate.
-     * @returns The certificate template.
-     */
-  static _generateCertificateTemplate(attr, years) {
-    let certificate = forge.pki.createCertificate();
-    certificate.setSubject(attr);
-    certificate.setIssuer(attr);
-    certificate.validity.notBefore = new Date();
-    certificate.validity.notAfter.setFullYear(
-      certificate.validity.notBefore.getFullYear() + years);
-    return certificate;
-  }
-
-  /**
-     * Generate a CA certificate.
-     * @param {forge.pki.PublicKey} publicKey The public Key of the CA.
-     * @param {forge.pki.PrivateKey} privateKey The private Key of the CA.
-     * @param {Object} certificateSubjects The object defining the content of CSR subjects.
-     * @param {Number} years The valid time interval of the generated certificate. The default value is 1.
-     * @returns A CA certificate.
-     */
-  static _generateCACertificate(publicKey, privateKey, certificateSubjects, years=1) {
-    let attrs = this._formattedSubjects(certificateSubjects);
-    let caCertificate = this._generateCertificateTemplate(attrs, years);
-    caCertificate.publicKey = publicKey;
-    caCertificate.serialNumber = '01';
-    caCertificate.setExtensions([{
-      name: 'basicConstraints',
-      cA: true,
-    }, {
-      name: 'subjectKeyIdentifier',
-    }, {
-      name: 'authorityKeyIdentifier',
-      keyIdentifier: true,
-    }]);
-    caCertificate.sign(privateKey);
-    return caCertificate;
-  }
-
-  /**
-     * Generate the verification certificate.
-     * @param {forge.pki.PrivateKey} caPrivateKey The CA private key to sign the verification certificate.
-     * @param {forge.pki.Certificate} caCertificate The CA certificate to provide the CSR subjects.
-     * @param {forge.pki.KeyPair} verificationKeys The key pair for verification.
-     * @param {Number} years The valid time interval of the generated certificate. The default value is 1.
-     * @returns A verification certificate.
-     */
-  static _generateVerificationCertificate(caPrivateKey, caCertificate, verificationKeys, years=1) {
-    let attrs = caCertificate.subject.attributes;
-    let certificate = this._generateCertificateTemplate(attrs, years);
-    certificate.publicKey = verificationKeys.publicKey;
-    certificate.sign(caPrivateKey, forge.md.sha256.create());
-    return certificate;
-  }
-
-  /**
-     * Convert the property object into the desired format.
-     * @param {Object} props The property object defining the content of CSR subjects.
-     * @returns An array with formatted subjects.
-     */
-  static _formattedSubjects(props) {
-    return [{
-      name: 'commonName',
-      value: props.commonName || '',
-    }, {
-      name: 'countryName',
-      value: props.countryName || '',
-    }, {
-      shortName: 'ST',
-      value: props.stateName || '',
-    }, {
-      name: 'localityName',
-      value: props.localityName || '',
-    }, {
-      name: 'organizationName',
-      value: props.organizationName || '',
-    }, {
-      shortName: 'OU',
-      value: props.organizationUnitName || '',
-    }];
-  }
-
-  /**
-     * Get the cetificates for registering a CA.
-     * The returned object contains the public key of the CA,
-     * the private key of the CA,
-     * the certificate of the CA,
-     * the public key of the verification,
-     * the private key of the verification,
-     * and the certificate of the verification.
-     * @param {Object} caCertSubjects The object defining the content of CSR subjects.
-     * @returns
-     */
-   static getCaRegistrationCertificates(caCertSubjects) {
-    const caKeys = forge.pki.rsa.generateKeyPair(2048);
-    const caCertificate = this._generateCACertificate(
-      caKeys.publicKey, caKeys.privateKey, caCertSubjects);
-    const verificationKeys = forge.pki.rsa.generateKeyPair(2048);
-    const verificationCertificate = this._generateVerificationCertificate(
-      caKeys.privateKey, caCertificate, verificationKeys);
-    const certificates = {
-      ca: {
-        keys: {
-          publicKey: forge.pki.publicKeyToPem(caKeys.publicKey),
-          privateKey: forge.pki.privateKeyToPem(caKeys.privateKey),
-        },
-        certificate: forge.pki.certificateToPem(caCertificate),
-      },
-      verification: {
-        keys: {
-          publicKey: forge.pki.publicKeyToPem(verificationKeys.publicKey),
-          privateKey: forge.pki.privateKeyToPem(verificationKeys.privateKey),
-        },
-        certificate: forge.pki.certificateToPem(verificationCertificate),
-      },
-    };
-    return certificates;
-  }
-};
