--- src/lambda-assets/ca-registrator/errors.js
@@ -1,9 +1,0 @@
-exports.UnknownVerifierError = class UnknownVerifierError extends Error{
-    static code = 422;
-    constructor() {
-        super("Received unknown verifier");
-    }
-    get code() {
-        return UnknownVerifierError.code;
-    }
-}
